Struct elastiql::aggregation::Aggregation

pub struct Aggregation {
    pub name: String,
    pub avg: Option<InnerAggregation>,
    pub weighted_avg: Option<WeightedAverageAggregation>,
    pub cardinality: Option<InnerAggregation>,
    pub max: Option<InnerAggregation>,
    pub min: Option<InnerAggregation>,
    pub median_absolute_deviation: Option<InnerAggregation>,
    pub percentiles: Option<InnerAggregation>,
    pub percentile_ranks: Option<InnerAggregation>,
    pub stats: Option<InnerAggregation>,
    pub extended_stats: Option<InnerAggregation>,
    pub sum: Option<InnerAggregation>,
    pub value_count: Option<InnerAggregation>,
    pub filters: Option<CompoundQuery>,
    pub terms: Option<InnerAggregation>,
    pub range: Option<RangeAggregation>,
    pub date_range: Option<DateRangeAggregation>,
    pub date_histogram: Option<DateHistogramAggregation>,
    pub auto_date_histogram: Option<AutoDateHistogramAggregation>,
    pub bucket_script: Option<BucketScript>,
    pub bucket_selector: Option<BucketSelector>,
    pub bucket_sort: Option<BucketSort>,
    pub nested: Option<NestedAggregation>,
    pub reverse_nested: Option<ReverseNestedAggregation>,
    pub metadata: Option<Map>,
    pub aggregations: Option<Vec<Aggregation>>,
}

An aggregation can be seen as a unit-of-work that builds analytic information over a set of documents.

Fields
name: String

The name for this aggregation.

NOTE: this must be unique otherwise the query will only return results for the last one with this name.

avg: Option<InnerAggregation>

A single-value metrics aggregation that computes the average of numeric values that are extracted from the aggregated documents. These values can be extracted either from specific numeric fields in the documents, or be generated by a provided script.

weighted_avg: Option<WeightedAverageAggregation>

A single-value metrics aggregation that computes the weighted average of numeric values that are extracted from the aggregated documents. These values can be extracted either from specific numeric fields in the documents.

When calculating a regular average, each datapoint has an equal "weight"... it contributes equally to the final value. Weighted averages, on the other hand, weight each datapoint differently. The amount that each datapoint contributes to the final value is extracted from the document, or provided by a script.

As a formula, a weighted average is the ∑(value * weight) / ∑(weight)

A regular average can be thought of as a weighted average where every value has an implicit weight of 1.

cardinality: Option<InnerAggregation>

A single-value metrics aggregation that calculates an approximate count of distinct values. Values can be extracted either from specific fields in the document or generated by a script.

max: Option<InnerAggregation>

A single-value metrics aggregation that keeps track and returns the maximum value among the numeric values extracted from the aggregated documents.

min: Option<InnerAggregation>

A single-value metrics aggregation that keeps track and returns the minimum value among numeric values extracted from the aggregated documents.

median_absolute_deviation: Option<InnerAggregation>

This single-value metrics aggregation approximates the median absolute deviation of its search results.

Median absolute deviation is a measure of variability. It is a robust statistic, meaning that it is useful for describing data that may have outliers, or may not be normally distributed. For such data it can be more descriptive than standard deviation.

It is calculated as the median of each data point’s deviation from the median of the entire sample. That is, for a random variable X, the median absolute deviation is median(|median(X) - Xi|).

percentiles: Option<InnerAggregation>

A multi-value metrics aggregation that calculates one or more percentiles over numeric values extracted from the aggregated documents.

Percentiles show the point at which a certain percentage of observed values occur. For example, the 95th percentile is the value which is greater than 95% of the observed values.

Percentiles are often used to find outliers. In normal distributions, the 0.13th and 99.87th percentiles represents three standard deviations from the mean. Any data which falls outside three standard deviations is often considered an anomaly.

When a range of percentiles are retrieved, they can be used to estimate the data distribution and determine if the data is skewed, bimodal, etc.

Note: Percentiles are (usually) estimated.

percentile_ranks: Option<InnerAggregation>

A multi-value metrics aggregation that calculates one or more percentile ranks over numeric values extracted from the aggregated documents.

Note: Percentiles are (usually) estimated.

stats: Option<InnerAggregation>

A multi-value metrics aggregation that computes stats over numeric values extracted from the aggregated documents.

The stats that are returned consist of: min, max, sum, count and avg.

extended_stats: Option<InnerAggregation>

A multi-value metrics aggregation that computes stats over numeric values extracted from the aggregated documents.

The extended_stats aggregations is an extended version of the stats aggregation, where additional metrics are added such as sum_of_squares, variance, std_deviation and std_deviation_bounds.

sum: Option<InnerAggregation>

A single-value metrics aggregation that sums up numeric values that are extracted from the aggregated documents.

value_count: Option<InnerAggregation>

A single-value metrics aggregation that counts the number of values that are extracted from the aggregated documents. These values can be extracted either from specific fields in the documents, or be generated by a provided script. Typically, this aggregator will be used in conjunction with other single-value aggregations.

filters: Option<CompoundQuery>

A bucketing aggregation that creates a single bucket for all documents matching the filters specified.

terms: Option<InnerAggregation>

A multi-bucketing value source based aggregation where buckets are dynamically built - one per unique value.

range: Option<RangeAggregation>

A range (bucketing) aggregation.

date_range: Option<DateRangeAggregation>

A range (bucketing) aggregation that is dedicated for date values and supports Date Math expressions.

date_histogram: Option<DateHistogramAggregation>

A multi-bucketing aggregation similar to the normal histogram aggregation, but can only be used with date or date range values.

auto_date_histogram: Option<AutoDateHistogramAggregation>

A multi-bucket aggregation similar to the Date histogram aggregation except instead of providing an interval to use as the width of each bucket, a target number of buckets is provided indicating the number of buckets needed and the interval of the buckets is automatically chosen to best achieve that target. The number of buckets returned will always be less than or equal to this target number.

bucket_script: Option<BucketScript>

A parent pipeline aggregation which executes a script which can perform per bucket computations on specified metrics in the parent multi-bucket aggregation. The specified metric must be numeric and the script must return a numeric value.

bucket_selector: Option<BucketSelector>

A parent pipeline aggregation which executes a script which determines whether the current bucket will be retained in the parent multi-bucket aggregation. The specified metric must be numeric and the script must return a boolean value.

bucket_sort: Option<BucketSort>

A parent pipeline aggregation which sorts the buckets of its parent multi-bucket aggregation. Zero or more sort fields may be specified together with the corresponding sort order. Each bucket may be sorted based on its _key, _count or its sub-aggregations. In addition, parameters from and size may be set in order to truncate the result buckets.

nested: Option<NestedAggregation>

A special single bucketing aggregation that enables aggregating nested documents.

Note: must be used with the aggregations field as a sibling.

reverse_nested: Option<ReverseNestedAggregation>

A special single bucketing aggregation that enables aggregating on parent docs from nested documents. Effectively this aggregation can break out of the nested block structure and link to other nested structures or the root document, which allows nesting other aggregations that aren’t part of the nested object in a nested aggregation.

The ReverseNestedAggregation aggregation must be defined inside a nested aggregation.

metadata: Option<Map>

The custom metadata to associate with this aggregation that will be returned alongside the results for this aggregation.

If the optional special key _skip is set to true, the results for this aggregation will be calculated but not returned with the other results.

aggregations: Option<Vec<Aggregation>>

The sub aggregation, if any.

Implementations

impl Aggregation

pub fn builder(
) -> AggregationBuilder<((), (), (), (), (), (), (), (), (), (), (), (), (), (), (), (), (), (), (), (), (), (), (), (), (), ())>

Create a builder for building Aggregation. On the builder, call .name(...), .avg(...)(optional), .weighted_avg(...)(optional), .cardinality(...)(optional), .max(...)(optional), .min(...)(optional), .median_absolute_deviation(...)(optional), .percentiles(...)(optional), .percentile_ranks(...)(optional), .stats(...)(optional), .extended_stats(...)(optional), .sum(...)(optional), .value_count(...)(optional), .filters(...)(optional), .terms(...)(optional), .range(...)(optional), .date_range(...)(optional), .date_histogram(...)(optional), .auto_date_histogram(...)(optional), .bucket_script(...)(optional), .bucket_selector(...)(optional), .bucket_sort(...)(optional), .nested(...)(optional), .reverse_nested(...)(optional), .metadata(...)(optional), .aggregations(...)(optional) to set the values of the fields (they accept Into values). Finally, call .build() to create the instance of Aggregation.