Module elastiql::search::query

Elasticsearch Query DSL types.

Structs

BooleanQuery
A query that matches documents matching boolean combinations of other queries. It is built using one or more boolean clauses, each clause with a typed occurrence.

CompoundQuery
Compound queries wrap other compound or leaf queries, either to combine their results and scores, to change their behavior, or to switch from query to filter context.

ExistsQuery
An Exists query returns documents that contain a non-null or empty (e.g. []) value for a field.

MatchQuery
A Match query returns documents that match a provided text, number, date or boolean value. The provided text is analyzed before matching.

NestedQuery
A Nested query wraps another query to search nested fields.

Query
A single search query.

QueryField
Describes a field that can be queried and its type.

QueryStringQuery
Query string returns documents based on a provided query string, using a parser with a strict syntax.

RangeQuery
A Range query returns documents that contain terms within a provided range.

RegexpQuery
A Regexp query returns documents that contain terms matching a regular expression.

SimpleQueryStringQuery
A Simple query string returns documents based on a provided query string, using a parser with a limited but fault-tolerant syntax.

TermQuery
A Term query returns documents that contain an exact term in a provided field.

TermsQuery
A Terms query returns documents that contain one or more exact terms in a provided field.