Struct elastiql::search::Cursor

pub struct Cursor(_);

An opaque cursor (obtained from the cursor field on an Edge) to return results after. This facilitates results pagination according to the Relay Cursor Connections Specification.

Note: it is expected that this will be used in conjunction with the same sort argument that created this Cursor.

Implementations

impl Cursor

pub fn encode(&self) -> Result<String, Box<dyn Error + Send + Sync>>

Encodes this Cursor to a base64 JSON string.

pub fn is_empty(&self) -> bool

Returns true if the Cursor contains no elements.